Run server tests in parallel: one mocha worker per core
Third PR of the server-test speedup series (after #2818 and #2820). With
the suite dominated by real test code, mocha's --parallel mode spreads
the files over one worker process per core (4 on the CI runners).

Measured locally on a 4-vCPU machine: full suite 73s -> ~37s, with the
same pass/fail list as the serial run.

How the pieces fit:

- test/bootstrap.test.js becomes test/hooks.js, a mocha root hook plugin
  loaded through --require: hooks declared inside a test file do not
  apply across files in parallel mode. Each worker boots its own Gladys
  (memoized: mocha calls beforeAll once per FILE, workers are reused for
  many files), and the API server listens on port 0 — no test depends on
  the actual port, supertest wraps the app object.
- setup-env.js derives a per-process SQLITE_FILE_PATH (the DuckDB file
  and the reset snapshot both follow automatically, being derived from
  that path), and removes the worker's database files at exit.
- nock is required first in hooks.js: it patches the core http module at
  load time, and each service embeds its own axios whose follow-redirects
  captures http.request when required — the patch must come first
  regardless of the file order a worker happens to load.
- The eslint devDependencies allowance is widened to test helper files
  (test/**/*.js), since hooks.js is no longer named *.test.js.
- npm run test-serial keeps the sequential mode available for debugging.
