+++ server/.eslintrc.json
@@ -75,7 +75,7 @@
     "import/no-extraneous-dependencies": [
       "error",
       {
-        "devDependencies": ["**/*.test.js"]
+        "devDependencies": ["**/*.test.js", "test/**/*.js"]
       }
     ],
     "max-len": [
