+++ server/package.json
@@ -8,8 +8,9 @@
   },
   "scripts": {
     "postinstall": "node ./cli/install_service_dependencies.js",
-    "test": "cross-env NODE_ENV=test SQLITE_FILE_PATH=./gladys-test.db ./node_modules/mocha/bin/mocha --require ./test/setup-env.js --recursive ./test/bootstrap.test.js \"./test/**/*.test.js\" --exit",
-    "test-service": "cross-env NODE_ENV=test SQLITE_FILE_PATH=/tmp/gladys-test.db ./node_modules/mocha/bin/mocha --require ./test/setup-env.js --recursive ./test/bootstrap.test.js \"./test/services/$npm_config_service/*.test.js\" \"./test/services/$npm_config_service/**/*.test.js\" --exit",
+    "test": "cross-env NODE_ENV=test SQLITE_FILE_PATH=./gladys-test.db ./node_modules/mocha/bin/mocha --require ./test/setup-env.js --require ./test/hooks.js --parallel --recursive \"./test/**/*.test.js\" --exit",
+    "test-serial": "cross-env NODE_ENV=test SQLITE_FILE_PATH=./gladys-test.db ./node_modules/mocha/bin/mocha --require ./test/setup-env.js --require ./test/hooks.js --recursive \"./test/**/*.test.js\" --exit",
+    "test-service": "cross-env NODE_ENV=test SQLITE_FILE_PATH=/tmp/gladys-test.db ./node_modules/mocha/bin/mocha --require ./test/setup-env.js --require ./test/hooks.js \"./test/services/$npm_config_service/*.test.js\" \"./test/services/$npm_config_service/**/*.test.js\" --exit",
     "coverage": "c8 npm test && c8 report --reporter=text-lcov > coverage.lcov",
     "prettier-check": "prettier --check '**/*.js' '**/*.json'",
     "prettier": "prettier --write '**/*.js' '**/*.json'",
